ocr_service/pyproject.toml
[project]
name = "ocr-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "paddleocr[doc-parser]>=3.4.0",
    "paddlepaddle-gpu==3.2.0",
    "python-multipart>=0.0.22",
    "uvicorn>=0.40.0",
    "numpy<2.0.0",
    "pydantic-settings>=2.12.0",
    "python-json-logger>=4.0.0",
    "httpx>=0.28.1",
]

[tool.uv.sources]
paddlepaddle-gpu = [
    { index = "paddlepaddle-gpu" }
]

[[tool.uv.index]]
name = "paddlepaddle-gpu"
url = "https://www.paddlepaddle.org.cn/packages/stable/cu118/"
explicit = true

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "ruff>=0.14.14",
]

[tool.mypy]
plugins = ["pydantic.mypy"]
ignore_missing_imports = true
# strict = true
# no_implicit_reexport = false

[tool.ruff.lint]
ignore = ["E501"]
extend-select = ["E", "I", "F", "G", "N", "T201", "T203"]
